UpdatePerson renamed to generic Update in mapper + plate repository
