+++ backend/WebApplication1/Models/OpenALPR/Responses/PlateResponse.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebApplication1.Models.Response
+namespace WebApplication1.Models.OpenALPR.Responses
 {
     public class PlateResponse
     {
